Abstract partial implementation of Parser interface added. Code duplication will be reduced when new parsers are added.
